move g2 related functions from utils to g2
